Update func_802300F0 interface from PR#83

+++ src/kernel/code_30EA0.h
@@ -3,6 +3,6 @@
 
 #include <PR/ultratypes.h>
 
-s32 func_802300F0(u16);
+u16 func_802300F0(s32);
 
 #endif // KERNEL_CODE_30EA0_H
